added prop to avoid delete of user when not having correct extension in user email

+++ src/main/java/no/vigo/provisioning/qlik/QlikUserService.java
@@ -84,7 +84,7 @@ public class QlikUserService {
                             addNeededGroups(hasGroups, neededGroups, user.id);
                             removeNotNeededGroups(hasGroups, neededGroups, user.id);
                         }
-                    } else {
+                    } else if (props.getUserDelete()){
                         log.info("Deleting user {}", user.mail);
                         userService.deleteUser(user.id);
